Add ADMINISTER command to vacuum.
Calling `administer vacuum(SomeType, OtherType.ptr, full := true)` will
vacuum the table for `SomeType` and all its descendants, as well as the
link table associated with the `ptr` (assuming that `ptr` is multi) and
all its descendants. The named only argument `full` being set to `true`
indicates the `FULL` option.

All of these arguments can be omitted. Not specifying any types or
pointers to vacuum will result in vacuuming everything that is
accessible to the user.

Co-authored-by: Michael J. Sullivan <sully@msully.net>

Fixes #6663