Fix spurious delete/create of constraints with arguments in migrations
The bug here was that `@index` was not populated in the `params` link
for concrete constraints by the reflection writer, and so when we
reloaded the schema from the database, it came back in an arbitrary
order.
This bug has been present for a *long* time, as far as I can tell, but
we tended to get "lucky" with the order pg returned the params.

This is a strictly *forward* fix. It can be cherry-picked but it will
only fix things for constraints created once the fix is applied. I'm
working on a backward fix as well.

Fixes #6699.