Fix issues with empty sets leaking out of optional scopes
For the previous issues, see #5575 and #6526.

The issue here is that in queries like
```
select User filter .avatar ?= <Card>{} and .name = 'Bob'
```
we were returning an empty set, since we were deciding that the
optional wraper we grabbed `.avatar` in was also a reasonable source
for User itself. We have machinery to prevent that (introduced in
the mask of the statement that was a "tentative container" for the
relation, but in this case, that container statement was not being
used, so the masks had no effect.

Instead, set up the masks on the final main rvar before including it.

Fixes #6739.