Branches, and fast branch creation (#6721)
The idea is to use pg_dump. This actually requires invoking pg_dump a few times, and making some tweaks to the SQL before executing it, in order to get all the pieces in. This is so that we can apply it on top of the template.

I also tried just doing a full dump/restore onto an empty database, which was much simpler to implement but was a lot slower.