Make DROP DATABASE properly wait for connections to drop
DROP DATABASE attempts to wait in a retry loop for all incoming
connections to be dropped, but it doesn't quite work, and so we have
hacked around it with another retry loop in the test suite.

Turns out the problem was that we were looking for the *edgedb*
database name in pg_stat_activity, not the *postgres* database name.

Fixes #4567.