Make DROP DATABASE properly wait for connections to drop (#6782)
DROP DATABASE attempts to wait in a retry loop for all incoming
connections to be dropped, but it doesn't quite work, and so we have
hacked around it with another retry loop in the test suite.

One problem was that we were looking for the *edgedb* database name in
pg_stat_activity, not the *postgres* database name.

The other source of issues was that, as I suspected in #4567, the
connection pool was establishing new connections to the database we
were trying to close. This came in two forms:
 * Transfers scheduled *after* we closed all the connections (since
   now the db was under quota!): this we fix by setting a flag to
   prevent.
 * Transfers that are *already* in flight. This we fix by waiting for
   all of the pending connections to complete, and then closing them
   too.

Once those issues were fixed, there were some failures in the test
suite where DROP failed because of remaining *edgedb* connections.
The problem there turned out to be in the test suite, where we were
cancelling some connections which resulted in non-synchronous closes.

Fixes #4567.