IDEA: Introduce a WITH based syntax for FOR-style iteration
I dislike how `for` looks and feels so much "heavier" than `with`, due
to the `union` keyword and the need to parenthesize the body if it is
a query, and especially if you need to have *nested* `for`s.

So my idea is to introduce `<-` binding, which behaves like a `for`
binding.

So then you can do something like
```
with name <- {'Sully', 'Zhibo', 'Aljaž'}
insert Person { name := name }
```
I modified some tests you can look at to compare.

Is this a bad idea, and I'm just sleep-deprived?
Does anybody but me like this?
(Does anybody have any other ideas about lighter-weight `for`?)