Support session and backend extension configs
This requires:
 * Some special handling for reflection, since extension settings
   don't show up in the views we look at normally
 * Updating _apply_session_config to work for extension configs
 * Making StateSerializerFactory able to include extension configs.

To test this, I fixed pg_trgm's configs. pg_trgm attempted to have
some backend configs, but the config object didn't extend
ExtensionConfig and also the backend_setting wasn't json.  I added
some checks to the extension path to prevent this sort of mistake.

Database configs work too, and I tested it manually, but currently the
only reliable way to make database configs with backend settings take
effect is to restart the server, so no tests for that yet.

The other big motivation for this is pgvector probes, which I'll leave
for @vpetrovykh to follow up on.