Partial rework of how lprop scope tree visibility works (#6775)
The core motivating thing here is that currently basically any use of
linkprops breaks the `[is Foo]` optimization where we only select from
`Foo`.
So if you have something like:
```
select schema::Constraint {
    ancestors[is schema::Constraint]: { @index },
};
```
it will select all `InheritingObject`s.

This is because `@index` is attached to the scopetree in a *fully*
flattened way, as
```
  Constraint.ancestors[is Constraint]@index
  Constraint.ancestors[is Constraint]
  Constraint.ancestors
  Constraint
```
because referring to a link property requires the source *of the link*
to be visible (to prevent the link from deduplicating), but this is
implemented by forcing the every path element to be visible. This means
that in the backend compiler, `Constraint.ancestors` is visible when we
compile the link property, which suppresses the optimization because
it looks like `Constraint.ancestors` may be getting used somewhere
*without* immediately type intersecting it.

Instead we try to do
```
  Constraint.ancestors[is Constraint]@index
    |-Constraint.ancestors[is Constraint]
      |- Constraint.ancestors
  Constraint
```

I think overall this is a code cleanup (_maybe_fixup_lprop can be
deleted), though I suspect there might still be a cleaner way to
structure the attach_path loop.

Some parts of relgen needed some fixing up in order to still be able
to find linkprop sources now that less is flattened.