Properly report errors involving newly created types
Currently, if we run a DDL command that errors, and we are not in a
transaction, we will ISE instead of interpreting the error correctly.
This is because the updated user schema doesn't get stored anywhere,
because we weren't in a transaction. When inside a transaction, it
would be stored in the dbview.

I originally fixed this by using implicit transactions to keep track
of the new user schema, but this required keeping implicit
transactions open on failures, and in opening up an implicit
transaction even in single statement execution of DDL, and that all
seemed questionable.

I wound up fixing it by stashing the schema in the exception object
and extracting it in interpret_error.

Fixes #6845.