Make CREATE EXTENSION ... VERSION more flexible
It will now select something >= the specified version while still
matching the major version.

In the future we can start adding qualifiers to the string to be more
or less specific.

Doing something like this is desired because existing migrations have
pgvector 0.4.0 written in them, which we'd prefer to not have exist in
edgedb 5.0.