Handle changes to vector length in migrations
We can't actually alter it, but we can suggest a DROP/CREATE.

If that change is rejected, we currently generate a topological sort
ISE. I still think this change is an improvement, though, and I'll be
following up with a separate thing about that ISE, that feels
conceptually quite separate.

(My plan there is that I think we should start reporting the cycle
errors as actual errors, since I think they frequently are real errors
that may be productively understood that way.)

Fixes #6854, partially.