Fix DML coalesce inside of IF/ELSE
The core thing here is that when we move a compiled IR expression into
a FOR loop, we need to adjust the scope tree so that it is inside the
loop's scope.

Also fix DML appearing directly as a FOR iterator, which I ran into
while testing a minimized bug (that turned out to not be the full
issue). I was nervous my change there might mess up the GROUP
optimizations, so I added another codegen test.

Fixes main part of #6794.