Refactor compiler pool a little
It seemed like too much boilerplate for each trivial function we added
to the compiler pool. Bring it down to 3 lines + 1 whitespace.

It would be trivial bring it down to 1 line by having a generic
function, but then the generic function's name would appear in
backtraces instead of the real name. We could dynamically generate
function bodies and get both, but that seemed like too much nonsense
for a problem that doesn't require a lot of nonsense.