Don't leak objects out of access policies when used in a computed global
The caching for our computed globals failed to account for whether
access policies are in effect, which could allow leaking of objects
that should not have been visible. Fix this by augmenting the cache
key.

A downside of this fix is that sometimes now we will generate two
identical computations of a computed global. This is probably fine for
now, since this is an important bug to fix, and 2 is still a lot
better than the N that it was without the cache.