Use generics to have better typing for Set
The idea here is that Set becomes SetE, which has a type param that
determines the type of the expr attribute. This allows us to refine
the type and have functions depend on expr's type.

Set itself becomes an alias for SetE[Expr | None].
Eventually I hope that it will be able to become an alias for
SetE[Expr]!