Fully eliminate Pointer.target/Pointer.target_path_id
Instead, pass the full Set in where it is needed in relctx.  This lets
us clean up some gunk but at the cost of requiring a Set in a lot of
places. (It was always trivial to change that, though).

I think it will be nicer once Pointer is an Expr and we can track it
in our type system with the new stuff.