Some test suite perf tweaks
Because of the new implementation of branches, creating databases from
a template is serialized now, which slows down test suites that rely
on creating database clones. Fortunately most of them don't really
need it, so we can simplify matters by avoiding branches for test
suites with empty databases.

Improve the rendering of tests a bit too by including a count of the
currently running tests and rerendering when a test *starts*, instead
of just when one finishes.