Start refactoring irast.Set (#6938)
* Use generics to have better typing for Set

The idea here is that Set becomes SetE, which has a type param that
determines the type of the expr attribute. This allows us to refine
the type and have functions depend on expr's type.

Set itself becomes an alias for SetE[Expr | None].
Eventually I hope that it will be able to become an alias for
SetE[Expr]!

* Introduce a TypeRoot Expr node to represent selecting from a type

This allows us to start reducing the ways in which the purpose of Set
is overloaded. The next step is moving Pointer from rptr to expr, and
then Set really will be just the expression metadata/scope tracker.

* Remove target field from irast.Pointer

I've always found this circularity suspicious. Just store the path_id.

* Fully eliminate Pointer.target/Pointer.target_path_id

Instead, pass the full Set in where it is needed in relctx.  This lets
us clean up some gunk but at the cost of requiring a Set in a lot of
places. (It was always trivial to change that, though).

I think it will be nicer once Pointer is an Expr and we can track it
in our type system with the new stuff.