Return JSON for magic link register
If the request specifies `Accept: application/json`, then even failures
should return JSON for the register request. `redirect_on_failure` is
still used to redirect in the case there is an issue when the user
follows the link since that's a GET request by a user agent like a
browser.