New docs review (#6992)
* Update installation method links

* Link to library docs instead of repos

* Update showcase links

* Add missing client libraries

* Remove redundant docs link

The user is already in the docs

* Refer to these by our name for them

Seems weirdly inconsistent to try to use a more generic name for access
policies here. We should call them by the name we call them. This will
otherwise be confusing, especially when a user clicks "Object-level
security" and lands on a page called "Access policies."

* Set correct link targets

These both pointed to the top of the "Select" page instead of to the
appropriate sections.

* Update to documentation links

instead of linking to the clients' GitHub repos

* Convert headings to proper HTML headings

* Add missing cheatsheets

* Mention Cloud in section about `EDGEDB_INSTANCE`

* Web -> web

* Fix link target for redeploying EdgeDB instance

* Add missing slash to showcase links

* Expand link label

Link goes to client connection configuration settings documentation
rather than client connection documentation. Better for accessibility to
expand the label to include "settings."

* Update deployment guide links

* Fix book link

* Update to reflect Elixir client is now blessed

* Add missing underscore

* Docs links updates (#6993)

* Update some docs links to use refs + add/remove host for some links to work with new docs site

* Fix CI

* Convert more client lib links to refs

* Add query parameter types heading ref

* Ignore missing ref errors for client docs

* Update docs/guides/datamigrations/postgres.rst

Co-authored-by: Aleksandra <alexsandra.sikora@gmail.com>

* Format library lists as proper lists

---------

Co-authored-by: James Clarke <james@edgedb.com>
Co-authored-by: Aleksandra <alexsandra.sikora@gmail.com>