Fix authenticate-related nightly test failure
A new test in #6996 relied on start_edgedb_server starting a server
using insecure_dev_mode, which was only true when the tests were run
from a dev-mode build.

Specify insecure_dev_mode explicitly in that test, and change the
default in start_edgedb_server to avoid this kind of problem in the
future.

(I actually don't know how to produce and run something locally that
works in non dev-mode.)