Add optional PKCE challenge in email verification
Currently, the email verification endpoint assumes
that if you only provide the email address, you
are not in a PKCE flow, which means the
verification ends with a redirect, or no content.
However, most of the supporting libraries expect
that verification ends with a PKCE code exchange.
Adding an optional PKCE challenge to the email
verification endpoint allows the client perform
the PKCE code exchange even if they only have an
email.