Skip creating @source/@target on derived views
This gave a 20-25% speedup on loading a big production schema.

In general I think that hugely slimming down how we create transient
types during compilation is the biggest opportunity for a big
performance improvement, and there are promising approaches that will
require some big changes. This gets a bunch done for very little.