Prevent parser from repeating unexpected errors
In unrecoverable cases, parser will generate "Unexpected token"
error for each of the tokens until EOF.

We discard all but the first error in python code, so this has
no practical effect at the moment. It will have it when we expose
all parser errors somewhere.

Also, this is a slight performance improvement, since we are not
allocating error structs (which I've seen exist in number of 15000+),
and passing them to Python only to be discarded.