Fix issues with cached global shapes and global cardinality inference (#7062)
Having an actual shape on a cached global caused confusion with the
convoluted logic used to determine whether something was a cached
global in range_for_material_objtype.
Instead, track it directly in TypeRoot.

Also, properly compile the type_rewrite for cached globals in a
detached scope so that the cardinality inference for it doesn't
collide with the enclosing scope.

Fixes #7056. Fixes #7061.