Drop special handling of type intersection in card inference
The goal of all that code is so that when you do things like `.<foo[is
Bar]`, if `Bar.foo` is an exclusive link then the card can be inferred
as `AT_MOST_ONE`. But setgen already restricts the backlink pointer
using the intersections in `resolve_ptr`, so I don't think this is
needed.