Add a debug flag to inject the edgeql as a comment in the SQL
This should make it easier to extract information from postgres logs
in certain situations.

Right now it is a debug flag named EDGEDB_DEBUG_PRESERVE_QUERIES.
Other options would be to do it unconditionally or to have a config
value.
Taking suggestions on names.

It injects the pre-constant extraction source.