Add support for marking indexes as `deferred`
A deferred index is a kind of index that gets created and updated in a
background process and not immediately.  This is required for indexes
that are slow and/or require calling out to external services.

To implement this we add the ability to specify `deferrability` on an
abstract index as one of three enum values:

* `Required`: concrete indexes of this type must be declared as
  `deferred`
* `Prohibited`: concrete indexes of this type must NOT be declared
  as `deferred`
* `Permitted`: either way is OK.

All current index types are marked as `Prohibited`.