Fix two issues directly reading pointers from a group
The first is that `select (group Card by .element).key.element`
produces an ISE. The issue here is that the path_id was registered at
the *use site* already in the path_scope. We already suppress the
path_scope for these things when compiling the group itself, and we
need to do it when compiling the result.
(Another scope tree related bug!)

The second is that in some cases, projecting a multi property out of a
shape, the the array that the multi property is serialized into could
escape and be returned from the query as the final result, which can
lead to an array being interpreted as a string, for example.

I actually fixed this in two ways:
 * First, shape compilation preents the shape elements from leaking out.
 * Second, we set `expr_exposed = False` when compiling the LHS
   of pointer dereferences,so we skip pointlessly compiling the
   shape also.

Fixes #5651. Fixes #6557.