Move extension casts into `__ext_casts__` module
Currently they are stored in `__derived__`, which is questionable and
requires some fiddly handling (see #7139).

I had originally wanted to place the casts in one of the modules that
the type is declared in, so that we can get rid of all the special
case cast deletions for extensions (and just delete everything in the
module), but I realized that won't work if we have an extension that
declares casts between a new type and a type in one of its
dependencies.