Fix two GROUP bugs involving USING clauses
There was a bug where we accidentally duplicated each user-written
alias in GROUP's USING clause. What's more, they were duplicated while
referring to the same AST node, which caused an incorrect cache hit on
the expr_view_cache, which led to not having a shape on the link in the output.
(I am skeptical that that cache still does anything
useful. I'll double check it.)

The other bug is that if you refer to a pointer directly but also have
a `USING` binding for that same name, the pointer reference will
actually refer to the using binding, not the pointer.

The only thing that made sense for this is to disallow this.
There's no way to distinguish them in key or grouping.

Fixes #4897.