Remove `id` property from `FreeObject` (#7126)
We do this by making FreeObject not extend BaseObject.

FreeObject ids are currently generated; this is not particularly
useful and requires some machinations to make work correctly.

Having FreeObject not have ids makes them more sensible, I think,
especially once we are going to support taking them as input.

This breaks compatibility but probably not in a way many people depend
on.

It might require some fixes to bindings. I fixed the python bindings a
while ago in prep for
this. (https://github.com/edgedb/edgedb-python/pull/464)