Provide infrastructure for long-running extension tasks
Some extensions might benefit from the ability to run an async task
whenever the extension is present in a branch.  Add a
`start_stop_extensions()` hook which is called whenever the set of
enabled extensions in a branch changes.

Additionally, make it possible to create named tasks associated with
the server tenant.