Implement support for `ext::ai::index`
The `ext::ai::index` is (currently) an always-deferred index.  Under the
hood it adds several new columns to the relation of the object type it
is declared on: one for each embedding vector variant (if the text
extraction model supports outputs of varying dimensionality, a.k.a
Matryoshka Representation), and one to denote if the embeddings are
up-to-date with respect to the object content (maintained by a trigger).

Declaration of `ext::ai::index` indexes also results in population of
several internal views that expose objects-to-be-indexed for the
deferred indexing process to consume.

The text extraction model used to convert the index expression to an
embedding is passed as a keyword argument when declaring a concrete
index.  Model metadata (name, dimensionality, limits etc) is then
persisted on the index as internal annotations for ease of access.