Fix stray loop variable use in _propagate_if_expr_refs
We were computing the "dummy" value based on whatever was left over in
the `fn` variable in an earlier loop, instead of doing it for the
field actually being considered. Sink that down into the loop.

This doesn't typically matter, but it's cleaner and it has an impact
for some experimental work I'm doing.