Fix initialization of defaults in patches (#7210)
In upgrade_schema, when we fix up schemas loaded from older versions,
we were filling new fields in with the defaults. That's not how the
schema is supposed to work, though, so fill them in with None instead.
(Defaults are resolved on the read side, not the write side.)

The particular issue being fixed here is with ObjectList, which is not
pickleable.

This was also probably the cause of the schema mismatch that required
a repair when we introduced pgvector. I don't think that fixing this
means we can skip the repair for the ai index changes, though, because
those also involve the `inherited` flag.