[5.x] Fix several bugs in the AI extension (#7220)
I'm making these changes on 5.x first and will forward port to
master. This is because I think that the patch mechanism is a critical
component.

 * Fix trivial index overriding by tweaking ddl_identity.
 * Make UPDATE triggers not always ISE
 * Filter out rows with NULL embeddings in fts::search
 * Update the triggers when overriding an index
 * Fix nontrivial index overriding.

The last one is the most impactful thing here.

The issue was that indexes in children were expected to have the same
name as in the parent, but did not. The "effective object index" in a
child, if the child defined the index itself, won't actually be a
child of the parent index, which was how the name was found.  It would
not be hard to do this computation right (and I had that part
working), but this approach has a major downside: *adding* an index to
a parent can require the column name to be changed in the child.
Additionally, it can't work if there are multiple independent base types
with ai indexes.

Another approach would be to always use the indexes *own* id as the
name.  I started along this path too, and had it much of the way. The
main downside (which I hadn't started on) is that handling the EXPLAIN
mode where inhviews are not used would be a pain.  This is *almost*
compatible with the existing database layouts (since nontrivial
inheritance doesn't work anyway) but differs when there is trivial
inheritance of indexes (where the child has no index specified).
So we would need to apply a patch to rename columns and whatnot.

Instead I went with a simplifying approach of making everything
use the same name. It's generalized in a way were it would be
very easy to use the model_name in the name instead, though.

This did also require some very custom patching code to rename all of
the generated columns, refresh the triggers, and refresh some views.
Refreshing the triggers was also necessary on is own, since they never
worked.

Co-authored-by: Victor Petrovykh <victor@edgedb.com>