Fix dropping AI indexes in a parent and child at the same time (#7227)
We need to skip putting objects that are in the process of being deleted
into the views, since the annotations may already be gone.

Also, the annotations that get copied are always marked OWNED, and we
need to preserve that when DROP OWNED.

Annoyingly, this requires passing context back in as an argument to
several functions that I just *removed* it as an argument from. When I
backport this, I'll need to make it optional.