Allow multiple authentication methods per transport in `--default-auth-method` (#7224)
Currently, `--default-auth-method` only allows one default method per
transport. This isn't very flexible and so allow multiple authentication
methods to be tried in sequence (according to the specified order in
`--default-auth-method`).

Technically, this infrastructure also allows trying multiple configured
`cfg::Auth` methods in order of `.priority`, but that is a bigger (and
breaking) change so I left it out.