Rebuild 4.0 dumps
I'm not sure /exactly/ which version the originals were built with.

The main motivation here is that test_dump_v4.py had FTS stuff
added *after* the dumps were generated and so skipped them.
We might update dump to actually include __fts_document__,
so we need to have dumps that *fail* to include it in order
to test that case.

Rebuild the dumps with a 4.0 release artifact but current 4.x dump
tests.

Done with:
```
sully@red:~/.local/share/edgedb/portable/4.0/lib/python3.11/site-packages
$ STABLE_DUMP=1 ~/.local/share/edgedb/portable/4.0/bin/python3 -m edb.tools --no-devmode gen-test-dumps -t ~/src/e/4.x-edgedb/tests/
```