Make ai::search have integrated sort and hit indexes
Tweak ai::search codegen to make it hit the index reliably even with
filtering NULLs out. It seems that postgres *can* sometimes manage to
use an ORDER BY index even when the function call isn't directly in
the ORDER BY, but it is much more fragile (broken by adding the NULL
check in #7223, for one).

Making ai::search return sorted output makes it easy to hit the
indexes and improves ergonomics.

Also:
 * Compile the arguments in the enclosing scope, which helps us hit the
   index in more complex scenarios (like a cast from json)
 * Make sure to export a source rvar for `.object`