Remove type simplification from schema. (#7228)
Make type simplification for schema utils `ensure_union_type` and `ensure_intersection_type` explicit. Remove the type simplification when generating types from a user schema.

When creating types from a schema, the implicit simplification can cause an error when the name of the simplified type differs from the non-simplified type. This can occur if the type of a link is the union of a type and a base type (eg. `type A; type A_ extending A;` `link foo -> A | A_`).