Fix doing a no-op update to an exclusive multi pointer with children
Currently, if a type has children and has a multi pointer with an
exclusive constraint, updating the pointer to the value it already has
will produce a ConstraintError. This is because while `UPDATE`
triggers check if the row has actually changed, updates on link tables
are done using deletes and inserts, and so the trigger function is
called unconditionally.

Fix this by checking for source in the trigger function.