Don't inject exclusive conflict checks for updates without children (#7271)
If we do two updates and the same type, then we currently inject
an exclusive conflict check even if there are no descendant types.
Avoid doing that, which also simplifies the check. (Though I was
nervous enough about the change that the expanded comment took
up all the deleted lines :).

This "fixes" the error in the example in #7263, though adding a
subtype of `B` makes it show up again. Real fix forthcoming.