Fix regression in using some tuple literals as a default
When we set defaults, we try to set a SQL default as well, since it
can greatly speed up populating a column.

If the naturally inferred type of the expression doesn't match the
property type exactly (say, `(0, 0)` getting inferred as `tuple<int64,
int64>` for a property with type `tuple<int32, int32>`), we need to
try casting it.

That cast gets expanded into something like
`(<int32>((0, 0).0), <int32>((0, 0).0))`.

That expression would get compiled to something postgres rejects.
Reject that sort of thing in the simple expression compiler path. This
also fixes some ISEs when an expression like this gets used in a
constraint or index.

The cause of the regression was that previously, `irutils.is_const`
would report tuple dereferences as not being const, even though they
clearly are. This change was introduced in #6938, the TypeRoot
refactor.